Final workaround for the Linux NFS file mapping problem.
When segments are unmapped we mark them as dirty and ready for writing
using the msync(MS_ASYNC) call. This in itself (along with munmap() and
even fclose()), does not guarantee that a disk write will occur. To ensure
this we need to also call fsync() before closing the file.